Refactor out more Lit components
