+++ src/components/BackChannelSidebar.ts
@@ -6,7 +6,10 @@
 
 import { LitElement, html, css, TemplateResult } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
-import type { IBackChannelPlugin, CaptureComment } from '../types';
+import type { IBackChannelPlugin } from '../types';
+import type { CommentsSection } from './CommentsSection';
+import './FeedbackForm.js';
+import './CommentsSection.js';
 
 /**
  * BackChannel Sidebar Component
@@ -21,12 +24,6 @@ export class BackChannelSidebar extends LitElement {
   visible: boolean = false;
 
   @state()
-  private comments: CaptureComment[] = [];
-
-  @state()
-  private loading: boolean = false;
-
-  @state()
   private showCommentForm: boolean = false;
 
   @state()
@@ -36,18 +33,6 @@ export class BackChannelSidebar extends LitElement {
     textContent: string;
     [key: string]: unknown;
   } | null = null;
-
-  @state()
-  private commentText: string = '';
-
-  @state()
-  private commentAuthor: string = '';
-
-  @state()
-  private isSubmitting: boolean = false;
-
-  @state()
-  private formError: string = '';
 
   static styles = css`
     :host {
@@ -150,82 +135,6 @@ export class BackChannelSidebar extends LitElement {
       padding: 20px;
     }
 
-    .comments-section {
-      margin-top: 20px;
-    }
-
-    .comments-title {
-      font-size: 16px;
-      font-weight: 600;
-      color: #333;
-      margin-bottom: 16px;
-    }
-
-    .comments-list {
-      display: flex;
-      flex-direction: column;
-      gap: 12px;
-    }
-
-    .comment-item {
-      background: #f8f9fa;
-      border-radius: 8px;
-      padding: 16px;
-      border-left: 4px solid #007acc;
-    }
-
-    .comment-meta {
-      font-size: 12px;
-      color: #666;
-      margin-bottom: 8px;
-    }
-
-    .comment-text {
-      font-size: 14px;
-      color: #333;
-      line-height: 1.4;
-    }
-
-    .comment-location {
-      font-size: 12px;
-      color: #666;
-      margin-top: 8px;
-      font-style: italic;
-    }
-
-    .empty-state {
-      text-align: center;
-      padding: 40px 20px;
-      color: #666;
-    }
-
-    .empty-state-icon {
-      font-size: 48px;
-      margin-bottom: 16px;
-      opacity: 0.5;
-    }
-
-    .loading {
-      text-align: center;
-      padding: 40px 20px;
-      color: #666;
-    }
-
-    .loading-spinner {
-      animation: spin 1s linear infinite;
-      font-size: 24px;
-      margin-bottom: 16px;
-    }
-
-    @keyframes spin {
-      0% {
-        transform: rotate(0deg);
-      }
-      100% {
-        transform: rotate(360deg);
-      }
-    }
-
     /* Responsive adjustments */
     @media (max-width: 768px) {
       :host {
@@ -267,170 +176,10 @@ export class BackChannelSidebar extends LitElement {
     .sidebar-content::-webkit-scrollbar-thumb:hover {
       background: #a1a1a1;
     }
-
-    /* Comment Form Styles */
-    .comment-form {
-      background: #f8f9fa;
-      border: 1px solid #dee2e6;
-      border-radius: 8px;
-      padding: 20px;
-      margin-bottom: 20px;
-    }
-
-    .comment-form-title {
-      font-size: 16px;
-      font-weight: 600;
-      color: #333;
-      margin-bottom: 16px;
-    }
-
-    .element-info {
-      background: #ffffff;
-      border: 1px solid #dee2e6;
-      border-radius: 6px;
-      padding: 12px;
-      margin-bottom: 16px;
-      font-size: 14px;
-      color: #666;
-    }
-
-    .element-info-label {
-      font-weight: 500;
-      color: #333;
-      margin-bottom: 4px;
-    }
-
-    .element-info-value {
-      font-family: monospace;
-      font-size: 12px;
-      color: #666;
-      word-break: break-all;
-    }
-
-    .form-group {
-      margin-bottom: 16px;
-    }
-
-    .form-label {
-      display: block;
-      font-size: 14px;
-      font-weight: 500;
-      color: #333;
-      margin-bottom: 6px;
-    }
-
-    .form-input {
-      width: 100%;
-      padding: 8px 12px;
-      border: 1px solid #ced4da;
-      border-radius: 4px;
-      font-size: 14px;
-      font-family: inherit;
-      transition: border-color 0.2s ease;
-    }
-
-    .form-input:focus {
-      outline: none;
-      border-color: #007acc;
-      box-shadow: 0 0 0 2px rgba(0, 122, 204, 0.2);
-    }
-
-    .form-textarea {
-      width: 100%;
-      padding: 12px;
-      border: 1px solid #ced4da;
-      border-radius: 4px;
-      font-size: 14px;
-      font-family: inherit;
-      min-height: 80px;
-      resize: vertical;
-      transition: border-color 0.2s ease;
-    }
-
-    .form-textarea:focus {
-      outline: none;
-      border-color: #007acc;
-      box-shadow: 0 0 0 2px rgba(0, 122, 204, 0.2);
-    }
-
-    .form-buttons {
-      display: flex;
-      gap: 12px;
-      justify-content: flex-end;
-    }
-
-    .form-button {
-      padding: 10px 20px;
-      border: none;
-      border-radius: 4px;
-      font-size: 14px;
-      font-weight: 500;
-      cursor: pointer;
-      transition: all 0.2s ease;
-    }
-
-    .form-button.primary {
-      background: #007acc;
-      color: white;
-    }
-
-    .form-button.primary:hover {
-      background: #0056b3;
-    }
-
-    .form-button.primary:disabled {
-      background: #6c757d;
-      cursor: not-allowed;
-    }
-
-    .form-button.secondary {
-      background: #6c757d;
-      color: white;
-    }
-
-    .form-button.secondary:hover {
-      background: #5a6268;
-    }
-
-    .form-error {
-      color: #dc3545;
-      font-size: 14px;
-      margin-top: 8px;
-      padding: 8px;
-      background: #f8d7da;
-      border: 1px solid #f5c6cb;
-      border-radius: 4px;
-    }
-
-    .form-success {
-      color: #155724;
-      font-size: 14px;
-      margin-top: 8px;
-      padding: 8px;
-      background: #d4edda;
-      border: 1px solid #c3e6cb;
-      border-radius: 4px;
-    }
-
-    .character-count {
-      font-size: 12px;
-      color: #666;
-      margin-top: 4px;
-      text-align: right;
-    }
-
-    .character-count.warning {
-      color: #ffc107;
-    }
-
-    .character-count.error {
-      color: #dc3545;
-    }
   `;
 
   connectedCallback() {
     super.connectedCallback();
-    this.loadComments();
     this.restoreVisibilityState();
   }
 
@@ -465,90 +214,11 @@ export class BackChannelSidebar extends LitElement {
 
       <div class="sidebar-content">
         ${this.showCommentForm ? this.renderCommentForm() : ''}
-        <div class="comments-section">
-          <h3 class="comments-title">Comments</h3>
-          ${this.renderComments()}
-        </div>
+        <comments-section
+          .backChannelPlugin=${this.backChannelPlugin}
+        ></comments-section>
       </div>
     `;
-  }
-
-  private renderComments(): TemplateResult {
-    if (this.loading) {
-      return html`
-        <div class="loading">
-          <div class="loading-spinner">⟳</div>
-          <div>Loading comments...</div>
-        </div>
-      `;
-    }
-
-    if (this.comments.length === 0) {
-      return html`
-        <div class="empty-state">
-          <div class="empty-state-icon">💬</div>
-          <div>No comments yet</div>
-          <div style="margin-top: 8px; font-size: 14px;">
-            Click "Capture Feedback" to add your first comment
-          </div>
-        </div>
-      `;
-    }
-
-    return html`
-      <div class="comments-list">
-        ${this.comments.map(comment => this.renderComment(comment))}
-      </div>
-    `;
-  }
-
-  private renderComment(comment: CaptureComment): TemplateResult {
-    const date = new Date(comment.timestamp).toLocaleString();
-    const elementHint = this.getElementHint(comment.location);
-
-    return html`
-      <div class="comment-item">
-        <div class="comment-meta">
-          ${comment.author ? `${comment.author} • ` : ''}${date}
-        </div>
-        <div class="comment-text">${comment.text}</div>
-        ${comment.snippet
-          ? html`<div class="comment-location">"${comment.snippet}"</div>`
-          : ''}
-        <div class="comment-location">${elementHint}</div>
-      </div>
-    `;
-  }
-
-  private getElementHint(xpath: string): string {
-    const parts = xpath.split('/');
-    const lastPart = parts[parts.length - 1];
-    if (lastPart.includes('[')) {
-      const tag = lastPart.split('[')[0];
-      return `${tag} element`;
-    }
-    return lastPart || 'page element';
-  }
-
-  private async loadComments(): Promise<void> {
-    if (!this.backChannelPlugin) return;
-
-    this.loading = true;
-    try {
-      const dbService = await this.backChannelPlugin.getDatabaseService();
-      const currentUrl = window.location.href;
-
-      // Get all comments and filter by current page URL
-      const allComments = await dbService.getComments();
-      this.comments = allComments.filter(
-        comment => comment.pageUrl === currentUrl
-      );
-    } catch (error) {
-      console.error('Failed to load comments:', error);
-      this.comments = [];
-    } finally {
-      this.loading = false;
-    }
   }
 
   private closeSidebar(): void {
@@ -603,7 +273,6 @@ export class BackChannelSidebar extends LitElement {
     this.visible = true;
     this.setAttribute('visible', 'true');
     this.updateVisibilityState();
-    this.loadComments();
   }
 
   /**
@@ -630,7 +299,12 @@ export class BackChannelSidebar extends LitElement {
    * Refresh the comments list
    */
   refreshComments(): void {
-    this.loadComments();
+    const commentsSection = this.shadowRoot?.querySelector(
+      'comments-section'
+    ) as CommentsSection;
+    if (commentsSection) {
+      commentsSection.refreshComments();
+    }
   }
 
   /**
@@ -644,9 +318,6 @@ export class BackChannelSidebar extends LitElement {
   }): void {
     this.selectedElement = elementInfo;
     this.showCommentForm = true;
-    this.commentText = '';
-    this.commentAuthor = '';
-    this.formError = '';
     this.requestUpdate();
   }
 
@@ -656,203 +327,37 @@ export class BackChannelSidebar extends LitElement {
   hideCommentForm(): void {
     this.showCommentForm = false;
     this.selectedElement = null;
-    this.commentText = '';
-    this.commentAuthor = '';
-    this.formError = '';
     this.requestUpdate();
   }
 
   private renderCommentForm(): TemplateResult {
     if (!this.selectedElement) return html``;
 
-    const textLength = this.commentText.length;
-    const maxLength = 1000;
-    const warningThreshold = 800;
-
-    let characterCountClass = 'character-count';
-    if (textLength > maxLength) {
-      characterCountClass += ' error';
-    } else if (textLength > warningThreshold) {
-      characterCountClass += ' warning';
-    }
-
     return html`
-      <div class="comment-form">
-        <h3 class="comment-form-title">Add Comment</h3>
-
-        <div class="element-info">
-          <div class="element-info-label">Selected Element:</div>
-          <div class="element-info-value">${this.selectedElement.tagName}</div>
-          ${this.selectedElement.textContent
-            ? html`
-                <div class="element-info-label" style="margin-top: 8px;">
-                  Content:
-                </div>
-                <div class="element-info-value">
-                  ${this.selectedElement.textContent.length > 100
-                    ? this.selectedElement.textContent.substring(0, 100) + '...'
-                    : this.selectedElement.textContent}
-                </div>
-              `
-            : ''}
-        </div>
-
-        <div class="form-group">
-          <label class="form-label" for="comment-text">Comment *</label>
-          <textarea
-            id="comment-text"
-            class="form-textarea"
-            placeholder="Enter your comment about this element..."
-            .value=${this.commentText}
-            @input=${this.handleCommentTextChange}
-            maxlength="1000"
-            required
-            ?disabled=${this.isSubmitting}
-          ></textarea>
-          <div class="${characterCountClass}">
-            ${textLength}/${maxLength} characters
-          </div>
-        </div>
-
-        <div class="form-group">
-          <label class="form-label" for="comment-author"
-            >Your Name (optional)</label
-          >
-          <input
-            id="comment-author"
-            type="text"
-            class="form-input"
-            placeholder="Enter your name or initials..."
-            .value=${this.commentAuthor}
-            @input=${this.handleCommentAuthorChange}
-            maxlength="100"
-            ?disabled=${this.isSubmitting}
-          />
-        </div>
-
-        ${this.formError
-          ? html`<div class="form-error">${this.formError}</div>`
-          : ''}
-
-        <div class="form-buttons">
-          <button
-            class="form-button secondary"
-            @click=${this.handleCancelComment}
-            ?disabled=${this.isSubmitting}
-          >
-            Cancel
-          </button>
-          <button
-            class="form-button primary"
-            @click=${this.handleSubmitComment}
-            ?disabled=${this.isSubmitting ||
-            !this.commentText.trim() ||
-            this.commentText.length > maxLength}
-          >
-            ${this.isSubmitting ? 'Saving...' : 'Save Comment'}
-          </button>
-        </div>
-      </div>
+      <feedback-form
+        .backChannelPlugin=${this.backChannelPlugin}
+        .selectedElement=${this.selectedElement}
+        @form-cancel=${this.handleFormCancel}
+        @comment-saved=${this.handleCommentSaved}
+      ></feedback-form>
     `;
   }
 
-  private handleCommentTextChange(event: Event): void {
-    const target = event.target as HTMLTextAreaElement;
-    this.commentText = target.value;
-    this.formError = '';
-    this.requestUpdate(); // Force re-render to update character count class
-  }
-
-  private handleCommentAuthorChange(event: Event): void {
-    const target = event.target as HTMLInputElement;
-    this.commentAuthor = target.value;
-  }
-
-  private handleCancelComment(): void {
+  private handleFormCancel(): void {
     this.hideCommentForm();
   }
 
-  private async handleSubmitComment(): Promise<void> {
-    if (!this.selectedElement || !this.commentText.trim()) {
-      this.formError = 'Please enter a comment.';
-      return;
-    }
+  private handleCommentSaved(event: CustomEvent): void {
+    const { comment, element } = event.detail;
 
-    if (this.commentText.length > 1000) {
-      this.formError = 'Comment is too long. Maximum 1000 characters allowed.';
-      return;
-    }
+    this.hideCommentForm();
+    this.refreshComments();
 
-    this.isSubmitting = true;
-    this.formError = '';
-
-    try {
-      const dbService = await this.backChannelPlugin.getDatabaseService();
-
-      // Store the selected element before clearing the form
-      const selectedElementInfo = this.selectedElement;
-
-      const comment = {
-        id: Date.now().toString(),
-        text: this.commentText.trim(),
-        pageUrl: window.location.href,
-        timestamp: new Date().toISOString(),
-        location: selectedElementInfo!.xpath,
-        snippet:
-          selectedElementInfo!.textContent?.substring(0, 100) || undefined,
-        author: this.commentAuthor.trim() || undefined,
-      };
-
-      await dbService.addComment(comment);
-
-      // Show success feedback
-      this.showSuccessMessage('Comment saved successfully!');
-
-      // Clear form and hide it
-      this.hideCommentForm();
-
-      // Refresh comments list
-      this.refreshComments();
-
-      // Notify parent about new comment for visual feedback
-      this.dispatchEvent(
-        new CustomEvent('comment-added', {
-          detail: { comment, element: selectedElementInfo },
-          bubbles: true,
-        })
-      );
-    } catch (error) {
-      console.error('Failed to save comment:', error);
-      this.formError = 'Failed to save comment. Please try again.';
-    } finally {
-      this.isSubmitting = false;
-    }
-  }
-
-  private showSuccessMessage(message: string): void {
-    const successDiv = document.createElement('div');
-    successDiv.className = 'form-success';
-    successDiv.textContent = message;
-    successDiv.style.cssText = `
-      position: fixed;
-      top: 20px;
-      right: 20px;
-      z-index: 10001;
-      background: #d4edda;
-      color: #155724;
-      border: 1px solid #c3e6cb;
-      border-radius: 4px;
-      padding: 12px 20px;
-      box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-      animation: slideIn 0.3s ease;
-    `;
-
-    document.body.appendChild(successDiv);
-
-    setTimeout(() => {
-      if (successDiv.parentNode) {
-        successDiv.parentNode.removeChild(successDiv);
-      }
-    }, 3000);
+    this.dispatchEvent(
+      new CustomEvent('comment-added', {
+        detail: { comment, element },
+        bubbles: true,
+      })
+    );
   }
 }
